core.go core_test.go new_connector.go unsupported_service_error.go
