--- src/hosting/core_test.go
@@ -1,1 +1,0 @@
-package hosting_test
