+++ src/hosting/unsupported_service_error.go
@@ -1,0 +1,14 @@
+package hosting
+
+import "errors"
+
+// UnsupportedServiceError communicates that the origin remote runs an unknown code hosting service.
+func UnsupportedServiceError() error {
+	return errors.New(`unsupported hosting service
+
+This command requires hosting on one of these services:
+* Bitbucket
+* GitHub
+* GitLab
+* Gitea`)
+}
